+++ junos/provider.go
@@ -150,6 +150,7 @@ func Provider() *schema.Provider {
 			"junos_system_login_user":                                    resourceSystemLoginUser(),
 			"junos_system_ntp_server":                                    resourceSystemNtpServer(),
 			"junos_system_radius_server":                                 resourceSystemRadiusServer(),
+			"junos_system_root_authentication":                           resourceSystemRootAuthentication(),
 			"junos_system_syslog_host":                                   resourceSystemSyslogHost(),
 			"junos_system_syslog_file":                                   resourceSystemSyslogFile(),
 			"junos_vlan":                                                 resourceVlan(),