+++ CHANGELOG.md
@@ -3,6 +3,7 @@ ENHANCEMENTS:
 * add `login` argument in `junos_system` resource (Fixes parts of [#88](https://github.com/jeremmfr/terraform-provider-junos/issues/88))
 * add `junos_system_login_class` resource (Fixes parts of [#88](https://github.com/jeremmfr/terraform-provider-junos/issues/88))
 * add `junos_system_login_user` resource (Fixes parts of [#88](https://github.com/jeremmfr/terraform-provider-junos/issues/88))
+* add `junos_system_root_authentication` resource
 
 BUG FIXES: